refactor: getKoreaTime 함수 리팩토링
utils 디렉토리를 없애고 logger 파일 내로 이동
