--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export { default as getKoreaTime } from './getKoreaTime';
-export * from './';
